requirements.txt
sentence-transformers
scikit-learn
faiss-cpu
numpy
pandas
tqdm
scikit-fuzzy
fastapi
uvicorn
matplotlib